requirements.txt
speechrecognition
pyaudio
openai
pillow
python-dotenv